// src/app/models/empcontributionmodel.ts
export interface empcontributionmodel{
    antsReviewComments: string;
    antsReviewRating: string;
    beyondTheScopeOfMywork: string;
    beyondTheScopeOfMyworkOne : string;
    beyondTheScopeOfMyworkTwo : string;
    employeeId: string;
    helpedMe: string;
    iHelped: string;
    periodId: string;

    
    carrerGoal: string;
    learningGoal: string;
    learningMeasure: string;    
    supportRequired: string;
    readinessForRole: string;
    roleEnrichManager: string;
    roleEnrichSelf: string;
    roleReadinessManager: string;
    roleReadinessSelf: string;

   

    antsImprovementsStart: string;
    antsImprovementsStop: string;
    antsImprovementsContinue: string;
    selfImprovementsStart: string;
    selfImprovementsStop: string;
    selfImprovementsContinue: string;
    userId: string;
  
}